refactor: 도메인별 config 및 job 디렉토리로 파일 구조 변경
