+++ module-batch/src/main/java/com/trybe/modulebatch/post/config/PostLikeJobConfig.java
@@ -1,4 +1,4 @@
-package com.trybe.modulebatch.config;
+package com.trybe.modulebatch.post.config;
 
 import org.springframework.batch.core.Job;
 import org.springframework.batch.core.Step;
