+++ module-batch/src/main/java/com/trybe/modulebatch/post/job/reader/PostLikeInsertReader.java
@@ -1,19 +1,18 @@
-package com.trybe.modulebatch.job.reader;
+package com.trybe.modulebatch.post.job.reader;
 
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Component;
 
 @Component
-public class PostLikeDeleteReader extends PostLikeReader {
-    public PostLikeDeleteReader(RedisTemplate<String, Long> redisTemplate) {
+public class PostLikeInsertReader extends PostLikeReader {
+    public PostLikeInsertReader(RedisTemplate<String, Long> redisTemplate) {
         super(redisTemplate);
     }
 
-    private final String REDIS_SCAN_PATTERN = "post:*:liked:users:deleted";
+    private final String REDIS_SCAN_PATTERN = "post:*:liked:users";
 
     @Override
     protected String getRedisScanPattern() {
         return REDIS_SCAN_PATTERN;
     }
-
 }
